Создал контроллер и страницу поиска
