+++ log/output.log
@@ -9,3 +9,7 @@
 2018-12-08 | 12:33:16 | INFO  | Database:121 | Candidates add sucsessfully | 
 2018-12-08 | 12:38:57 | INFO  | WebSecurityConfig:40 | user has successfully logged in  | 
 2018-12-08 | 12:38:58 | INFO  | MlHhApplication:39 | Add new candidates | 
+2018-12-09 | 12:05:58 | INFO  | WebSecurityConfig:40 | user has successfully logged in  | 
+2018-12-09 | 12:05:59 | INFO  | MlHhApplication:39 | Add new candidates | 
+2018-12-09 | 12:14:41 | INFO  | WebSecurityConfig:40 | user has successfully logged in  | 
+2018-12-09 | 12:14:42 | INFO  | MlHhApplication:39 | Add new candidates |
