+++ src/main/resources/templates/searchdb.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <title>Search page</title>
+</head>
+<body>
+<h1>Search page</h1>
+<img th:src="@{TacoCloud.png}"/>
+</body>
+</html>
